+++ st2common/st2common/content/loader.py
@@ -36,6 +36,9 @@ class ContentPackLoader(object):
     Class for loading pack and pack content information from directories on disk.
     """
 
+    # TODO: Rename "get_content" methods since they don't actually return
+    # content - they just return a path
+
     ALLOWED_CONTENT_TYPES = ['sensors', 'actions', 'rules', 'aliases', 'policies']
 
     def get_packs(self, base_dirs):
@@ -186,7 +189,7 @@ class ContentPackLoader(object):
     def _get_folder(self, pack_dir, content_type):
         path = os.path.join(pack_dir, content_type)
         if not os.path.isdir(path):
-            raise ValueError('No %s found in "%s".' % (content_type, pack_dir))
+            return None
         return path